# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: OpenPlan MVP1 API
  version: 1.0.0
  description: |
    재정합 정본 (US 07-15 기준). 봉투: 성공 {data, meta} / 실패 {error{code,message,details}}.
    오류 코드 전수 목록은 exceptions.md. x-user-stories = US/SS 추적, x-lane = W5 소유 레인.
    인증: httpOnly 쿠키 op_at(JWT access 30m) + op_rt(refresh 14d, Path=/api/v1/auth). ADR-0001.

    ── D-16 인증 구축 순서 (ADR-0010) ──
    인증(authN) 실구현은 4주차. 그 전에는 dev-auth 스텁이 고정 개발 사용자를 주입한다.
    x-auth-phase: WEEK4-STUB = 4주차까지 501 {error.code: E-AUTH-011} 스텁 /
    DEV-STUB-ACTIVE = 스텁 기간에도 dev 사용자 기준으로 동작 (FE 세션 복원 경로 보존).
    본 파일의 인증 요청/응답 스키마는 그 자체로 확정 계약이다 — 4주차 교체 시 계약 변경 0.
    dev 프로파일: devAuth(X-Dev-User 헤더, 생략 시 고정 dev 사용자) / 운영: cookieAuth.
    상세 스텁 계약: api-contracts.md §4.
servers:
  - url: /api/v1
security:
  - cookieAuth: []
  - devAuth: []   # dev 프로파일 한정 (D-16 dev-auth 스텁). 운영 빌드에서 필터 자체가 미등록
...
paths:
...
  /projects/{projectId}:
    get:
      tags: [projects]
      operationId: getProject
      summary: 프로젝트 상세 (태스크 상태 집계 배지 포함)
      x-lane: BE-2
      x-user-stories: [PROJ-03, PROJ-04]
      parameters: [{ $ref: "#/components/parameters/ProjectId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [projects]
      operationId: updateProject
      summary: 프로젝트 편집 (version 필수 — SYS-05)
      description: >-
        전체 교체다(부분 갱신 아님). 편집 가능 4필드(name·description·dueDate·priority)를 모두 실어
        보내야 하며, 생략한 필드는 null로 교체된다. status·closedAt은 편집으로 못 바꾼다(포함 시 400).
      x-lane: BE-2
      x-user-stories: [PROJ-06]
      parameters: [{ $ref: "#/components/parameters/ProjectId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version]
              properties:
                name: { type: string }
                description: { type: string }
                dueDate: { type: [string, "null"], format: date }
                priority: { type: [integer, "null"], minimum: 1, maximum: 3 }
                version: { type: integer, format: int64 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Project" }
        "409": { $ref: "#/components/responses/VersionConflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
    delete:
      tags: [projects]
      operationId: deleteProject
      summary: 삭제 — 태스크·계획 항목·수행 이력 일괄 제거 (NFR-008 일관 기준)
      x-lane: BE-2
      x-user-stories: [PROJ-09]
      parameters: [{ $ref: "#/components/parameters/ProjectId" }]
      responses:
        "204": { description: 삭제 완료 }
...
  /projects/{projectId}/duplication-preview:
    get:
      tags: [projects]
      operationId: getDuplicationPreview
      summary: "복제 항목 확인 (이름·설명·태스크·WBS 개요 — PROJ-11)"
      x-lane: BE-2
      x-user-stories: [PROJ-10, PROJ-11]
      parameters: [{ $ref: "#/components/parameters/ProjectId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      name: { type: string }
                      description: { type: string }
                      taskCount: { type: integer }
                      wbsItemCount: { type: integer }
                      note: { type: string, description: "주간 계획 항목은 복제되지 않음(47행 열거 밖) — 복제본 태스크 전량 미배치" }
        "404": { $ref: "#/components/responses/NotFound" }
  /projects/{projectId}/duplications:
    post:
      tags: [projects]
      operationId: duplicateProject
      summary: "복제 실행 (Idempotency-Key 수용 — 서버 dedup 아님, 아래 설명 참조)"
      x-lane: BE-2
      x-user-stories: [PROJ-12]
      parameters:
        - { $ref: "#/components/parameters/ProjectId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: false
        description: "본문 전체 생략 가능 — 생략(또는 {})하면 복제본 이름은 '원본명 (복제)'가 된다."
        content:
          application/json:
            schema:
              type: object
              # 기본 예시는 빈 객체 — Swagger UI가 newName 을 "string" 으로 자동 채워 보내
              # 복제본이 'string' 으로 생성되는 함정을 막는다.
              example: {}
              properties:
                newName:
                  type: string
                  maxLength: 100
                  description: "미지정·공백이면 '원본명 (복제)'(접미사 포함 100자 초과 시 원본명을 잘라 맞춤). 지정 시 생성과 동일 규칙(trim·1~100자, 위반 422)."
                  example: "마케팅 2026"
      responses:
        "201":
          description: >-
            복제본 생성 — 원본 무변경, 태스크 전량 UNASSIGNED, 복제본은 IN_PROGRESS.
            WBS 는 복제본 태스크로 재연결된다. 주간 계획 블록·수행 이력은 복사하지 않는다.


            🔴 Idempotency-Key 는 수용하지만 **서버가 dedup 하지 않는다** — 관측 로그로만 남긴다.
            같은 키로 두 번 POST 하면 복제본이 둘 생긴다(확정 POST /confirmation 과 다르다 — 그쪽은
            상태 기반 멱등이라 키 없이도 멱등이다). 이중 클릭 방지는 현재 **클라이언트 책임**이다.
            서버 dedup(키 저장소)은 전 도메인 공통 인프라라 후속 과제로 남아 있다.


            원본 마감일이 이미 지났으면 복제본 dueDate 는 null(무기한)이 된다 — 지난 마감일을
            그대로 물려주면 자동 종료(ADR-0006)가 다음 조회에서 복제본을 CLOSED 로 뒤집는다.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/Project" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: op_at
      description: 운영 모드 — httpOnly JWT access 쿠키 (ADR-0001). 4주차부터 활성.
    devAuth:
      type: apiKey
      in: header
      name: X-Dev-User
      description: |
        dev-auth 스텁 (D-16, ADR-0010). dev 프로파일에서만 필터 등록.
        헤더 생략 시 고정 dev 사용자(00000000-0000-0000-0000-000000000001)로 인증된 것으로 간주.
        시드된 다른 테스트 사용자 UUID를 넣으면 그 사용자로 동작 (NFR-030 사용자 격리 테스트용).
        미시드 UUID → 401 E-COM-002. 운영 프로파일에는 이 필터가 존재하지 않는다.
  parameters:
...
    IdempotencyKey: { name: Idempotency-Key, in: header, schema: { type: string, format: uuid } }
...
    ProjectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    NotFound:
      description: E-COM-004 리소스 없음 (타인 리소스 은닉 포함)
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    VersionConflict:
      description: E-COM-006 저장 충돌 (SYS-05) — details.latest에 최신본 동봉 (3택 UI 재료)
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
    UnprocessableEntity:
      description: E-COM-009 값 의미 오류 (5분 단위·시간 역전 등)
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } }
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, pattern: "^E-[A-Z]+-[0-9]{3}$" }
            message: { type: string, description: "사용자 표시 가능 문구 (P4 준수)" }
            details: { type: object, additionalProperties: true }
...
    ProjectSummary:
      type: object
      properties:
        projectId: { type: string, format: uuid }
        name: { type: string }
        dueDate: { type: [string, "null"], format: date }
        status: { type: string, enum: [IN_PROGRESS, PAUSED, CLOSED] }
        priority: { type: [integer, "null"], deprecated: true, description: "미사용 — 프로젝트 우선순위는 요구사항에 없다(FR-141·148). 항상 null. 코드 정리 후 삭제 예정" }
        badges:
          type: object
          x-implementation-status: not-implemented
          description: "[미구현] PROJ-01/04 배지 — 응답에 아직 없음. FE는 부재 시 0으로 폴백하므로 배지가 전부 0으로 보인다(조용한 실패). 텍스트 병기 (NFR-017)"
          properties:
            unassignedCount: { type: integer }
            assignedCount: { type: integer }
            deadlineSoon: { type: boolean }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
    Project:
      allOf:
        - { $ref: "#/components/schemas/ProjectSummary" }
        - type: object
          properties:
            description: { type: string }
            closedAt: { type: [string, "null"], format: date-time }
            taskStats:
              type: object
              x-implementation-status: not-implemented
              description: "[미구현] PROJ-04 태스크 상태별 개수 — 응답에 아직 없음. FE는 부재 시 0으로 폴백"
              properties:
                total: { type: integer }
                unassigned: { type: integer }
                inProgress: { type: integer }
                completed: { type: integer }
...